refactor parity scripts
